Extracts CoreData managed object models into separate files.
