+++ FeedModule/FeedModule/FeedCache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -57,50 +57,3 @@ public final class CoreDataFeedStore: FeedStore {
     }
     
 }
-
-// MARK: - Core Data object models
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-    @NSManaged var timestamp: Date
-    @NSManaged var feed: NSOrderedSet
-    
-    var localFeed: [LocalFeedImage] {
-        return feed.compactMap { ($0 as? ManagedFeedImage)?.local}
-    }
-    
-    func managedImageFeed(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        return NSOrderedSet(array: localFeed.map { localFeedImage in
-            let managedFeedImage = ManagedFeedImage(context: context)
-            managedFeedImage.id = localFeedImage.id
-            managedFeedImage.imageDescription = localFeedImage.description
-            managedFeedImage.location = localFeedImage.location
-            managedFeedImage.url = localFeedImage.url
-            return managedFeedImage
-        })
-    }
-    
-    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-        try find(in: context).map(context.delete)
-        return ManagedCache(context: context)
-    }
-    
-    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
-        request.returnsObjectsAsFaults = false
-        return try context.fetch(request).first
-    }
-    
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-    @NSManaged var id: UUID
-    @NSManaged var imageDescription: String?
-    @NSManaged var location: String?
-    @NSManaged var url: URL
-    @NSManaged var cache: ManagedCache
-    
-    var local: LocalFeedImage {
-        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
-    }
-}
