+++ FeedModule/FeedModule/FeedCache/Infrastructure/CoreData/ManagedFeedImage.swift
@@ -1,0 +1,21 @@
+//
+//  ManagedFeedImage.swift
+//  FeedModule
+//
+//  Created by Omran Khoja on 2/15/22.
+//
+
+import CoreData
+
+@objc(ManagedFeedImage)
+internal class ManagedFeedImage: NSManagedObject {
+    @NSManaged var id: UUID
+    @NSManaged var imageDescription: String?
+    @NSManaged var location: String?
+    @NSManaged var url: URL
+    @NSManaged var cache: ManagedCache
+    
+    var local: LocalFeedImage {
+        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
+    }
+}
